requirements.txt
appdirs==1.4.4
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==20.8b1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
pyparsing==3.0.7
pytz==2021.3
pyyaml
regex==2022.1.18
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
soupsieve==2.3.1
sqlparse==0.4.2
tenacity==8.0.1
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8